fix-demo

+++ global/page/page.go
@@ -9,17 +9,17 @@ const (
 )
 
 type Pagination struct {
-	CurrentPage uint `json:"current_page"` // 当前页码
-	PerPage     uint `json:"per_page"`     // 当前页行数
-	TotalPage   uint `json:"total_page"`   // 总页码
-	Total       uint `json:"total"`        // 总行数
+	CurrentPage int   `json:"current_page"` // 当前页码
+	PerPage     int   `json:"per_page"`     // 当前页行数
+	TotalPage   int   `json:"total_page"`   // 总页码
+	Total       int64 `json:"total"`        // 总行数
 }
 
-func GetPagination(page, pageSize uint, total uint) Pagination {
+func GetPagination(page, pageSize int, total int64) Pagination {
 	return Pagination{
 		Total:       total,
 		CurrentPage: page,
 		PerPage:     pageSize,
-		TotalPage:   uint(math.Ceil(float64(total) / float64(pageSize))),
+		TotalPage:   int(math.Ceil(float64(total) / float64(pageSize))),
 	}
 }
